bench: drop astype(f64) in stream reduction — bincount upcasts internally
np.bincount accumulates weights in float64 regardless of input dtype, so
passing the native f32/uint chunk.data directly is bit-identical to
astype(f64) first (verified max|d|=0 on hlca) while skipping a full-nnz
copy — ~5% faster end-to-end. Applied to the bench worker and the README
streaming example.

This is the shippable part of the lever-#1 overlap spike. The larger win
(overlapping decode with the reduce) needs a Rust-side fix — the producer
does not prefetch and the per-chunk PyBytes copy is GIL-bound, so a
Python prefetch thread gives no speedup. Findings + plan saved to
scratch/spike-overlap-findings.md.

+++ bench/python/runner.py
@@ -69,9 +69,10 @@ def run_stream(path: str, chunk_size: int) -> tuple[float, int, int]:
         # slow path): per-column weighted counts == per-gene sums. Keeping the
         # reduction cheap means wall_s reflects decode throughput, not the
         # reduction, so it is comparable to the read-dominated load path.
-        gene_sums += np.bincount(
-            chunk.indices, weights=chunk.data.astype(np.float64), minlength=n_vars
-        )
+        # Pass the native (f32/uint) weights directly — bincount accumulates in
+        # float64 internally, so the result is bit-identical to astype(f64)
+        # first, but skips a full-nnz copy (~5% faster end-to-end on hlca).
+        gene_sums += np.bincount(chunk.indices, weights=chunk.data, minlength=n_vars)
         n_obs += chunk.nrows
     total = float(gene_sums.sum()) if gene_sums is not None else 0.0
     return total, n_obs, n_vars
